+++ HISTORY.rst
@@ -5,6 +5,10 @@ History
 0.3 (unreleased)
 ----------------
 
+- Added 3DES authentication: authenticate()
+
+- Added get_card_uid()
+
 - Made unknown file types ignored, instead of raising an error